+++ example/bones_api_example.dart
@@ -56,7 +56,7 @@ class MyBTCModule extends APIModule {
   }
 }
 
-/// The `APIRoot` defines the version and modules to use:
+/// The `APIRoot` defines the API version and modules to use:
 class MyAPI extends APIRoot {
   MyAPI({dynamic apiConfig}) : super('example', '1.0', apiConfig: apiConfig);