+++ lib/src/bones_api_error_zone_generic.dart
@@ -1,0 +1,1 @@
+void printToZoneStderrImpl(Object? o) => print(o);