+++ lib/src/bones_api_error_zone_io.dart
@@ -1,0 +1,3 @@
+import 'dart:io';
+
+void printToZoneStderrImpl(Object? o) => stderr.writeln(o);